# statistics: add pool in the buildBucket4Merging to reuse datum