--- a/statistics/histogram.go
+++ b/statistics/histogram.go
@@ -101,6 +101,12 @@ type scalar struct {
 // EmptyScalarSize is the size of empty scalar.
 const EmptyScalarSize = int64(unsafe.Sizeof(scalar{}))
 
+var datumPool = sync.Pool{
+	New: func() any {
+		return &types.Datum{}
+	},
+}
+
 // NewHistogram creates a new histogram.
 func NewHistogram(id, ndv, nullCount int64, version uint64, tp *types.FieldType, bucketSize int, totColSize int64) *Histogram {
 	if tp.EvalType() == types.ETString {
@@ -128,10 +134,20 @@ func (hg *Histogram) GetLower(idx int) *types.Datum {
 	return &d
 }
 
+// GetLowerFromPool gets the lower bound of bucket `idx` from pool.
+func (hg *Histogram) GetLowerFromPool(idx int) *types.Datum {
+	return hg.Bounds.GetRowDatumFromPool(2*idx, 0, hg.Tp, &datumPool)
+}
+
 // GetUpper gets the upper bound of bucket `idx`.
 func (hg *Histogram) GetUpper(idx int) *types.Datum {
 	d := hg.Bounds.GetRow(2*idx+1).GetDatum(0, hg.Tp)
 	return &d
+}
+
+// GetUpperFromPool gets the upper bound of bucket `idx` from pool.
+func (hg *Histogram) GetUpperFromPool(idx int) *types.Datum {
+	return hg.Bounds.GetRowDatumFromPool(2*idx+1, 0, hg.Tp, &datumPool)
 }
 
 // MemoryUsage returns the total memory usage of this Histogram.
@@ -1141,8 +1157,12 @@ func (hg *Histogram) buildBucket4Merging() []*bucket4Merging {
 	buckets := make([]*bucket4Merging, 0, hg.Len())
 	for i := 0; i < hg.Len(); i++ {
 		b := newbucket4MergingForRecycle()
-		hg.GetLower(i).Copy(b.lower)
-		hg.GetUpper(i).Copy(b.upper)
+		lower := hg.GetLowerFromPool(i)
+		lower.Copy(b.lower)
+		datumPool.Put(lower)
+		upper := hg.GetUpperFromPool(i)
+		upper.Copy(b.upper)
+		datumPool.Put(upper)
 		b.Repeat = hg.Buckets[i].Repeat
 		b.NDV = hg.Buckets[i].NDV
 		b.Count = hg.Buckets[i].Count
@@ -1349,12 +1369,13 @@ func MergePartitionHist2GlobalHist(sc *stmtctx.StatementContext, hists []*Histog
 				minValue = hist.GetLower(0).Clone()
 				continue
 			}
-			res, err := hist.GetLower(0).Compare(sc, minValue, collate.GetBinaryCollator())
+			tmpValue := hist.GetLower(0)
+			res, err := tmpValue.Compare(sc, minValue, collate.GetBinaryCollator())
 			if err != nil {
 				return nil, err
 			}
 			if res < 0 {
-				minValue = hist.GetLower(0).Clone()
+				minValue = tmpValue.Clone()
 			}
 		}
 	}
@@ -1511,6 +1532,7 @@ func MergePartitionHist2GlobalHist(sc *stmtctx.StatementContext, hists []*Histog
 		}
 		globalHist.AppendBucketWithNDV(bucket.lower, bucket.upper, bucket.Count, bucket.Repeat, bucket.NDV)
 	}
+	datumPool.Put(minValue)
 	return globalHist, nil
 }