--- a/statistics/histogram.go
+++ b/statistics/histogram.go
@@ -128,8 +128,8 @@ func (hg *Histogram) GetLower(idx int) *types.Datum {
 	return &d
 }
 
-// GetLowerToDatum gets the lower bound of bucket `idx` to datum.
-func (hg *Histogram) GetLowerToDatum(idx int, d *types.Datum) {
+// LowerToDatum gets the lower bound of bucket `idx` to datum.
+func (hg *Histogram) LowerToDatum(idx int, d *types.Datum) {
 	hg.Bounds.GetRow(2*idx).DatumWithBuffer(0, hg.Tp, d)
 }
 
@@ -139,8 +139,8 @@ func (hg *Histogram) GetUpper(idx int) *types.Datum {
 	return &d
 }
 
-// GetUpperToDatum gets the upper bound of bucket `idx` to datum.
-func (hg *Histogram) GetUpperToDatum(idx int, d *types.Datum) {
+// UpperToDatum gets the upper bound of bucket `idx` to datum.
+func (hg *Histogram) UpperToDatum(idx int, d *types.Datum) {
 	hg.Bounds.GetRow(2*idx).DatumWithBuffer(0, hg.Tp, d)
 }
 
@@ -1151,8 +1151,8 @@ func (hg *Histogram) buildBucket4Merging() []*bucket4Merging {
 	buckets := make([]*bucket4Merging, 0, hg.Len())
 	for i := 0; i < hg.Len(); i++ {
 		b := newbucket4MergingForRecycle()
-		hg.GetLowerToDatum(i, b.lower)
-		hg.GetUpperToDatum(i, b.upper)
+		hg.LowerToDatum(i, b.lower)
+		hg.UpperToDatum(i, b.upper)
 		b.Repeat = hg.Buckets[i].Repeat
 		b.NDV = hg.Buckets[i].NDV
 		b.Count = hg.Buckets[i].Count