<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close https://github.com/pingcap/tidb/issues/46804

Problem Summary:

### What is changed and how it works?

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)


before: 

```
$  go test -run=^$ -bench=BenchmarkMergePartitionHist2GlobalHist -benchmem github.com/pingcap/tidb/statistics  

goos: darwin
goarch: arm64
pkg: github.com/pingcap/tidb/statistics
BenchmarkMergePartitionHist2GlobalHist/Size1000-8                     14          75695795 ns/op        45846930 B/op     782037 allocs/op
BenchmarkMergePartitionHist2GlobalHist/Size10000-8                     1        1669962708 ns/op        601507720 B/op  10201122 allocs/op
BenchmarkMergePartitionHist2GlobalHist/Size100000-8                    1        24380916125 ns/op       5989421424 B/op 101480169 allocs/op
PASS
ok      github.com/pingcap/tidb/statistics      30.015s
```

after:

```
$ go test -run=^$ -bench=BenchmarkMergePartitionHist2GlobalHist -benchmem github.com/pingcap/tidb/statistics
goos: darwin
goarch: arm64
pkg: github.com/pingcap/tidb/statistics
BenchmarkMergePartitionHist2GlobalHist/Size1000-8                     21          50100944 ns/op        26460268 B/op     381136 allocs/op
BenchmarkMergePartitionHist2GlobalHist/Size10000-8                     1        1330470875 ns/op        406708560 B/op   6142843 allocs/op
BenchmarkMergePartitionHist2GlobalHist/Size100000-8                    1        24889564666 ns/op       4066911992 B/op 61428000 allocs/op
PASS
ok      github.com/pingcap/tidb/statistics      30.540s


```


- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```